+++ NEWS.md
@@ -1,4 +1,4 @@
-# occumb (development version)
+# occumb v0.2.0
 * Change occumbData class specification and occumbData() function.
 * Add occumb() function.
 * Add methods for occumbFit class: plot, print, summary.